refactor: collapse ide barrel into ide/index.ts
